Fix clippy::useless_conversion in read_clock_ms
tv_nsec is already i64 on this target; the newer clippy on CI (unlike
the older toolchain used for local dev) flags the redundant
i64::from() wrap that a prior commit introduced.

+++ client/linux/src/capture.rs
@@ -396,7 +396,7 @@ fn read_clock_ms(clock_id: libc::clockid_t) -> CoreResult<i64> {
             std::io::Error::last_os_error().to_string(),
         ));
     }
-    Ok(ts.tv_sec * 1000 + i64::from(ts.tv_nsec) / 1_000_000)
+    Ok(ts.tv_sec * 1000 + ts.tv_nsec / 1_000_000)
 }
 
 impl LifecycleHooks for LinuxPlatformHooks {
